+++ src/qos.rs
@@ -522,7 +522,7 @@ pub struct EntityName {
 }
 
 unsafe fn user_data_from_qos_native(qos: *const dds_qos_t) -> Option<Vec<u8>> {
-    let mut sz: size_t = 0;
+    let mut sz: usize = 0;
     let mut value: *mut ::std::os::raw::c_void = std::ptr::null_mut();
     if dds_qget_userdata(qos, &mut value, &mut sz) {
         // Cyclone DDS returns a copy of the value so okay to take ownership
@@ -544,13 +544,13 @@ unsafe fn user_data_to_qos_native(qos: *mut dds_qos_t, user_data: &Option<Vec<u8
         dds_qset_userdata(
             qos,
             ptr as *const ::std::os::raw::c_void,
-            user_data.len() as size_t,
+            user_data.len(),
         );
     }
 }
 
 unsafe fn topic_data_from_qos_native(qos: *const dds_qos_t) -> Option<Vec<u8>> {
-    let mut sz: size_t = 0;
+    let mut sz: usize = 0;
     let mut value: *mut ::std::os::raw::c_void = std::ptr::null_mut();
     if dds_qget_topicdata(qos, &mut value, &mut sz) {
         // Cyclone DDS returns a copy of the value so okay to take ownership
@@ -572,13 +572,13 @@ unsafe fn topic_data_to_qos_native(qos: *mut dds_qos_t, topic_data: &Option<Vec<
         dds_qset_topicdata(
             qos,
             ptr as *const ::std::os::raw::c_void,
-            topic_data.len() as size_t,
+            topic_data.len(),
         );
     }
 }
 
 unsafe fn group_data_from_qos_native(qos: *const dds_qos_t) -> Option<Vec<u8>> {
-    let mut sz: size_t = 0;
+    let mut sz: usize = 0;
     let mut value: *mut ::std::os::raw::c_void = std::ptr::null_mut();
     if dds_qget_groupdata(qos, &mut value, &mut sz) {
         // Cyclone DDS returns a copy of the value so okay to take ownership
@@ -600,7 +600,7 @@ unsafe fn group_data_to_qos_native(qos: *mut dds_qos_t, group_data: &Option<Vec<
         dds_qset_groupdata(
             qos,
             ptr as *const ::std::os::raw::c_void,
-            group_data.len() as size_t,
+            group_data.len(),
         );
     }
 }
@@ -1321,14 +1321,14 @@ fn test_user_data_to_native() {
         let policy = Some(create_u8_vec_for_tests());
         user_data_to_qos_native(qos_native, &policy);
 
-        let mut sz: size_t = 0;
+        let mut sz: usize = 0;
         let mut value: *mut ::std::os::raw::c_void = std::ptr::null_mut();
         assert!(dds_qget_userdata(qos_native, &mut value, &mut sz));
 
         let output = Vec::from_raw_parts(
             value as *mut ::std::os::raw::c_uchar,
-            sz as usize,
-            sz as usize,
+            sz,
+            sz,
         );
         assert!(output.len() == 5);
         assert_eq!(output, policy.unwrap());
@@ -1347,7 +1347,7 @@ fn test_user_data_from_native() {
         dds_qset_userdata(
             qos_native,
             ptr as *const ::std::os::raw::c_void,
-            test_vec.len() as size_t,
+            test_vec.len(),
         );
 
         let policy = user_data_from_qos_native(qos_native);
@@ -1367,14 +1367,14 @@ fn test_topic_data_to_native() {
         let policy = Some(create_u8_vec_for_tests());
         topic_data_to_qos_native(qos_native, &policy);
 
-        let mut sz: size_t = 0;
+        let mut sz: usize = 0;
         let mut value: *mut ::std::os::raw::c_void = std::ptr::null_mut();
         assert!(dds_qget_topicdata(qos_native, &mut value, &mut sz));
 
         let output = Vec::from_raw_parts(
             value as *mut ::std::os::raw::c_uchar,
-            sz as usize,
-            sz as usize,
+            sz,
+            sz,
         );
         assert!(output.len() == 5);
         assert_eq!(output, policy.unwrap());
@@ -1393,7 +1393,7 @@ fn test_topic_data_from_native() {
         dds_qset_topicdata(
             qos_native,
             ptr as *const ::std::os::raw::c_void,
-            test_vec.len() as size_t,
+            test_vec.len(),
         );
 
         let policy = topic_data_from_qos_native(qos_native);
@@ -1413,14 +1413,14 @@ fn test_group_data_to_native() {
         let policy = Some(create_u8_vec_for_tests());
         group_data_to_qos_native(qos_native, &policy);
 
-        let mut sz: size_t = 0;
+        let mut sz: usize = 0;
         let mut value: *mut ::std::os::raw::c_void = std::ptr::null_mut();
         assert!(dds_qget_groupdata(qos_native, &mut value, &mut sz));
 
         let output = Vec::from_raw_parts(
             value as *mut ::std::os::raw::c_uchar,
-            sz as usize,
-            sz as usize,
+            sz,
+            sz,
         );
         assert!(output.len() == 5);
         assert_eq!(output, policy.unwrap());
@@ -1439,7 +1439,7 @@ fn test_group_data_from_native() {
         dds_qset_groupdata(
             qos_native,
             ptr as *const ::std::os::raw::c_void,
-            test_vec.len() as size_t,
+            test_vec.len(),
         );
 
         let policy = group_data_from_qos_native(qos_native);
